Moved files around
Moved directories into more natural places.
Corrected solution/project files.
Fixed some warnings for x86.

+++ AlphaLi.Test/util/static_for.test.cpp
@@ -1,5 +1,5 @@
 #include "stdafx.h"
-#include "util/static_for.h"
+#include "alphali/util/static_for.h"
 #include <array>
 #include <iostream>
 
@@ -14,14 +14,14 @@ TEST_CASE("Static for loop", "[util]") {
 		constexpr size_t size_loop = 4;
 		std::array<int, size_loop> arr = {0,0,0,0};
 
-		auto pred = [&arr](int i) {
-			arr[i] = i;
+		auto pred = [&arr](size_t i) {
+			arr[i] = static_cast<int>(i);
 		};
 
 		static_for<0, 4>()(pred);
 
-		for (auto i = 0; i < 4; ++i) CHECK(arr[i] == i);
-		static_for<0, 3>()([&arr](int i) { CHECK(arr[i] == i); });
+		for (auto i = 0; i < 4; ++i) CHECK(arr[static_cast<size_t>(i)] == i);
+		static_for<0, 3>()([&arr](int i) { CHECK(arr[static_cast<size_t>(i)] == i); });
 
 	}
